+++ CHANGELOG.md
@@ -4,6 +4,9 @@ All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [unreleased]
+- Move `newrelic_rpm.rb` mock to the `test/mocks` directory
+- Add `add_mocks_to_load_path`
+- Add NewRelicLogs formatter & appender
 
 - Add payload logging to `payload_includes` Minitest helper functionality.
 - Fix bug in `payload_includes` testing functionality where if a key is not in the payload the test would pass.